refactor: extract common logic to shared package
