+++ docs/docs/docs/standalone-cli.mdx
@@ -1,0 +1,25 @@
+import { PackageManagerTabs, Badge } from '@theme';
+
+# Standalone CLI
+
+## Installation
+
+To get started, install the package using your preferred package manager:
+
+<PackageManagerTabs command="install licenses-aggregator" />
+
+## Programmatic usage (API)
+
+Once the setup is complete, you can easily add a button to your app to display the list of open-source licenses:
+
+```tsx
+
+```
+
+This will open a native screen listing all detected licenses. The title of the screen can be customized by passing a different string argument to `launchLicenseListScreen()`.
+
+## Programmatic usage (API)
+
+## Additional details
+
+For more notes on the mechanics of the tool, please see [Quick Start](/docs/quick-start)
