+++ package.json
@@ -8,7 +8,7 @@
     ]
   },
   "scripts": {
-    "build:library": "yarn workspace react-native-legal build-plugins && yarn workspace react-native-legal build-library",
+    "build:library": "yarn workspace @callstack/react-native-legal-shared build && yarn workspace react-native-legal build-plugins && yarn workspace licenses-aggregator build-library && yarn workspace react-native-legal build-library",
     "lint:js": "eslint \"**/*.{js,ts,tsx}\"",
     "typescript": "yarn build:library && tsc --noEmit -p examples/bare-example/tsconfig.json && tsc --noEmit -p examples/expo-example/tsconfig.json",
     "lint:android": "./scripts/lint-android.sh",
@@ -39,6 +39,7 @@
     "@evilmartians/lefthook": "1.4.3",
     "commitlint": "17.6.6",
     "eslint": "^8.46.0",
+    "eslint-import-resolver-typescript": "^4.4.3",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-prettier": "^5.0.0",
     "prettier": "^3.0.1",
