+++ packages/licenses-aggregator/.editorconfig
@@ -1,0 +1,7 @@
+[*.{h,m,mm}]
+indent_style = space
+indent_size = 2
+
+[*.{swift}]
+indent_style = space
+indent_size = 2
