+++ packages/licenses-aggregator/README.md
@@ -1,0 +1,33 @@
+# licenses-aggregator
+
+Aggregate license notes of OSS libraries used in your Node.js project
+
+## Installation
+
+```sh
+npm install licenses-aggregator
+```
+
+## Usage
+
+
+```js
+import { multiply } from 'licenses-aggregator';
+
+// ...
+
+const result = await multiply(3, 7);
+```
+
+
+## Contributing
+
+See the [contributing guide](CONTRIBUTING.md) to learn how to contribute to the repository and the development workflow.
+
+## License
+
+MIT
+
+---
+
+Made with [create-react-native-library](https://github.com/callstack/react-native-builder-bob)
