+++ packages/licenses-aggregator/package.json
@@ -1,0 +1,67 @@
+{
+  "name": "licenses-aggregator",
+  "version": "1.0.0",
+  "description": "Aggregate license notes of OSS libraries used in your Node.js project",
+  "main": "build/index",
+  "module": "build/index",
+  "types": "build/index.d.ts",
+  "bin": {
+    "license-kit": "build/index.js"
+  },
+  "exports": {
+    ".": {
+      "source": "./src/index.ts",
+      "types": "./build/typescript/index.d.ts",
+      "default": "./build/module/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "src",
+    "build",
+    "!**/__tests__",
+    "!**/__fixtures__",
+    "!**/__mocks__",
+    "!**/.*"
+  ],
+  "scripts": {
+    "test": "jest",
+    "typecheck": "tsc",
+    "lint": "eslint \"**/*.{js,ts,tsx}\"",
+    "clean": "del-cli build",
+    "build-library": "tsc -p tsconfig.json"
+  },
+  "keywords": [
+    "nodejs",
+    "oss",
+    "license",
+    "acknowledgements",
+    "license-management"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/callstackincubator/react-native-legal.git",
+    "directory": "packages/licenses-aggregator"
+  },
+  "author": "Mateusz Mędrek <luigimario1913@gmail.com> (https://github.com/mateusz1913)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/callstackincubator/react-native-legal.git/issues"
+  },
+  "homepage": "https://callstackincubator.github.io/react-native-legal/",
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org/"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.5",
+    "@types/signale": "^1.4.7",
+    "del-cli": "^5.1.0",
+    "jest": "^29.7.0",
+    "typescript": "^5.8.3"
+  },
+  "dependencies": {
+    "@callstack/react-native-legal-shared": "workspace:*",
+    "commander": "^14.0.0",
+    "signale": "^1.4.0"
+  }
+}
