+++ packages/react-native-legal/package.json
@@ -35,7 +35,6 @@
     "ios",
     "lib",
     "plugin",
-    "plugin-utils",
     "src",
     "LICENSE",
     "app.plugin.js",
@@ -49,6 +48,7 @@
     "tsc": "tsc"
   },
   "devDependencies": {
+    "@types/glob": "^8.1.0",
     "@types/xml2js": "^0.4.11",
     "expo": "~52.0.36",
     "expo-module-scripts": "^4.0.0",
@@ -57,6 +57,7 @@
     "react-native-builder-bob": "0.23.1"
   },
   "dependencies": {
+    "@callstack/react-native-legal-shared": "workspace:*",
     "glob": "^7.1.3",
     "xcode": "^3.0.1",
     "xml2js": "^0.6.2"
