+++ packages/react-native-legal/plugin-utils/tsconfig.json
