+++ packages/shared/tsconfig.json
@@ -12,8 +12,7 @@
     "moduleResolution": "node",
 
     "outDir": "build",
-    "rootDir": "src",
-    "typeRoots": ["../types-definitions", "../../../node_modules/@types"]
+    "rootDir": "src"
   },
   "include": ["./src"],
   "exclude": ["**/__mocks__/*", "**/__tests__/*"]
